pyproject.toml
[project]
name = "market-entry-agent"
version = "0.1.0"
description = "Multi-agent system for international market entry research"
authors = [{name = "Christine Brown Clayton", email = "christine@tech.unieros.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.19.0",
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"